Refactor: Consolidate demo service and handlers
This commit consolidates the demo service logic into a single file and removes redundant handler files. It also updates the service implementation to use the new consolidated file.
